implement the logic for addresses in the account
took longer than i thought. the addresses could not be taken from the user that the userManager returns because it doesn't include any collections defined on the entity, this took me hours to realize. i debugged and literally saw addresses being added to the database for the user and when i was trying to get the addresses, it was an empty list. addresses had to be added using the ef core standard and queried for that way as well i.e. context.Addresses.AddAsync and context.Addresses.Where(...). when this was done, slowly everything else fell into place and now this functionality is on par with the original implementation