fix the post request for add new address
funny story, i tested the api from the swagger ui and it worked but i didn't realize that it didn't work from the client

two issues was discovered:
first, i received a 404 not found because i didn't set the uri correctly [missed the api/]
second, serializing to json and sending that as the content for the post request resulted in a 400 bad request where the CreateAddressModel was required and was driving me insane because i couldn't for the life of me figure out why then i took the json produced in the code and tested in swagger and it did the same thing, returned a 400 bad request. so now i'm getting somewhere. did some research and found jsoncontent.create and that worked like a charm