implement functionality to add tyres
some things had to change to make this work as expected and unfortunately, it has come at the cost of yet another modal. bootstrap modals don't play well with aspnetcore. since a file is uploaded and validated against, it is done on the server side so the add tyre modal had to be rewriiten as a view