remove a log message i used during debugging
what happened was that i had a property on the orderitemmodel called totalitemprice which was calculated based on the tyre price and quantity but it errored out when mapped from an order entity to an ordermodel because tyre was null even though it was being mapped. tried many ways to fix that but it would give so i removed it. seems like i left a log message behind