implement the viewing of orders on the account
ported over the html and cleaned it up quite a bit. removed useless styling and logic. now it's a much cleaner solution compared to the original

ran into a small issue however, when i originally specified the "on delete" action on the address in the order, i mistakenly specified a one-to-one relationship between the order and address which is inaccurate. one address could appear on many orders but one order cannot contain many addresses so that adjustment was made. what would happen is that the relationship would be severed after one order with the same address used on retrieving the orders from the database. it did continue to work when it is added to the database however