:D send an email when a customer completes an order
ported the email template and cleaned it up. now instead of it being generated through code [that's how i did it in the original which was horrible] a template is used and is embedded in the application

fluentemail is used to render the template and send the email

had an issue which [no joke] took five damn hours to figure out why the template wasn't working. there was a null reference in the ordermodel and i only figured out what when i gave up on the template and decided to create the email in code. once i resolved that, the template worked