implement the functionality to reset password
makes use of another email template but unfortunately, i couldn't get the spacing on certain clients to look right. firstly, the div tag is not respected in outlook, so a table is used instead. secondly, margin isn't respected as well so i gave up there, on gmail and samsung mail, it's respected and lastly, border-radius isn't respected in outlook too but it doesn't matter

also, sending the password reset token makes use of the built-in services to generate the token and verify it. in the original implementation, i manually generated that code and sent it and when the account password was reset, a new code and was generated and saved to the database. the current built-in services generates a huge token so it's much safer

yet another modal is removed here as well, leaving only one,  the viewreceiptmodal, which i'm happy to keep