microservices: add the last one, the address service
the webapi and data projects have been scrapped

the mvc project now holds the migrations for the customers database and is responsible for apiauthtoken generation since it's pointless having the authentication service endpoint. i'm aware that this is bad practice to have the client having so much of the responsiblility to authenticate users, however, this is just for the time being. later, identityserver4 will come into play and alleviate the client of these responsibilities and will also bring forth a backend-for-frontend to further alleviate responsibility from the client in terms of knowing the endpoints of the microservices and will return to just using one endpoint