implement an api gateway with ocelot
skills from Kevin Dockx's Securing Microservices in ASP.NET Core were utilized here

now all traffic is routed through the api gateway and allows for token exchange

before, issued access tokens were very permissive in that, it allowed access to all microservices and exposed the backend implementation. now that's not the case

the mvc frontend has a limited number of scopes it can ask for

the mvc bff [gateway] is now another api resource with a scope that the mvc frontend can request

when a user authenticates to the site, they only have access to the gateway and at the gateway, depending on the route, a dedicated delegating handler deals with the request performing a token exchange with identityserver4 to get the correct audience to access the microservices

an issue i had here was that the rsk.tokenexchange.identityserver4 package i had in place for token exchange wasn't working the way i had hoped so i had to use kevin's implementation for the token exchange extension grant validator to get it to work

this approach improves security by reducing the attack surface of the token should it be intercepted

the client credentials flow at the frontend still works but is adjusted to work with the gateway by requesting an additional scope, the gateway along with the required customer service scope to pass the policies at the identity / customer microservice. the gateway will perform its authentication and forward the request on

along the way i found a handy dandy little package called "Configuration Substitutor" that allows you to use reference variables in appsettings.json and have them substituted at runtime. this was needed for configuring ocelot since i wasn't happy with the idea of hardcoding routes to the microservices since it will become hard to maintain when the routes change

----

the seelanstyres.mvc project was also renamed to seelanstyres.frontends.mvc for consistency with the other projects