prep for centralized logging with serilog and elasticsearch
skills from two courses were utilized here:
- Implementing Cross-cutting Concerns for ASP.NET Core 3 Microservices by Steve Gordan
- Logging and Monitoring in ASP.NET Core 6 by Erik Dahl

since this is a prep, elasticsearch is not enabled by default

once the solution is containerized and orchestrated locally, then elasticsearch and kibana will come into play as containers

i've enriched the logs with the informational version from the assembly and custom metadata that may or may not be added to overtime

currently, a link to the commit in github is an extra piece of metadata in the assembly which will be useful in the logs to investigate exceptions

some az-400 skills also come into play here, notably:

- develop an instrumentation strategy
  - assess and configure a logging framework
  - design a log aggregation and storage strategy
  - design a log aggregation and query strategy
  - design and implement distributed tracing

for distributed tracing, aspnetcore sets the traceId when requests are made. in kibana or azure log analytics, querying by the traceId will provide all the logs relevant for a request if it fails and one can determine where the system failed to handle the request and why

those objectives mentioned earlier are from the az-400 that i studied for and took, the exam looks very different now

another note about those objectives, those objectives are done in code here. in the exam, for example, distributed tracing was tested using the application map of application insights and ofcourse, the projects have to be instrumented with application insights. i didn't do that here, not too sure how possible that is when the solution is in azure kubernetes services but i'll figure that out when i get there

in azure kubernetes services, the monitoring addon will be enabled [also from the exam]