the days and nights were long and so too, was the logging spree
my strategy for this was to start at the level of the microservices and work my way up to the frontend

structured logging was used everywhere with two handy fields to base alerts on:
- announcement
- stopwatchElapsedTime

announcements are everywhere and when something fails, the announcement is FAILED

stopwatchElapsedTime was created by timing certain tasks in the system. database calls, encryption, decryption, view model building to name a few

examples of possible actionable alerts:

for announcement:
- a query looking for announcement when it is set to FAILED and exceeds 10 in 3 minutes should raise an alert and notify developers so they can investigate the logs and track down the code that caused it. the commit url enrichment will be handy here, where new code is causing failures to occur

- for stopwatchElapsedTime
a query looking for stopwatchElapsedTime when it is exceeding 500ms at least 10 times in 3 minutes should trigger an alert notifying operations teams that the system is degraded and is responding slowly, indicating that the system needs to scale out to meet the demand. if that's not possible, i.e. it's on premises, tough luck but at least they will get notified and get to watch the system fail right in front of their eyes 😈🌩️⚡mwa⚡ha⚡ha⚡🌩️

motivation to move to azure i guess 😇

more az-400 skills are used / referenced here:
- develop a site reliability engineering strategy
  - implement alerts based on appropriate log messages
  - analyze a system with regards to load and failure conditions

again, these objectives are from the az-400 that i studied for and took. site reliability engineering is completely removed in the current version of the exam, wonder why microsoft did that but i'm sure it'll come back in some way, either as a specialty exam or added back slowly