containerize and MOSTLY orchestrate the solution locally
i say this with a heavy heart, even though i lost track of how many hoops i jumped through to the solution to this point, that i couldn't fully orchestrate the solution due to a rather stupid problem. host.docker.internal is used to enable communication between all the microservices but the identity service can not and does not want to work with that. i've tried various tactics over the passed two days to get it to work but it never did. no idea how microsoft got it to work in their eShopOnContainers reference architecture but oh well

in order to compensate for that disappointment, i tried my best to make the building and running of the solution as seamless as possible, at the cost of a single file publish. it's a small price to pay to make me happy. don't look in the publish folder, just stay in the clean scripts folder

there's various scripts that work together to build and run the solution which i explain well in the readme file so i won't mention it here but in summary, orchestration is simulated like it's running using docker compose

later when i eventually orchestrate with kubernetes and it gives the same problem, i'll have to deploy the identity service to azure app services

many changes were made to get to this point and i may miss some:
- removal of https. this was a nightmare for a day and caused most of the adjustments
- adjusting to use docker compose
- using a better way to enforce the port number, and that was to just configure kestrel 😐
- enriched the logs with the activityspan to include the traceId, it wasn't added by default
- since client configuration changes a lot, i decided to remove a client if found and add it again. saving me the trouble of always deleting the database and running the application again

i thought the last build will be the last to produce artifacts but i was wrong, it'll continue

in some rather good news, log aggregation, storage and querying as well as distributed tracing all work and work well. soon i'll add to the readme how to go about setting up and using kibana to see it in action