update the kibana demo to point to google drive
okay, so i uploaded the 1080p version to my google drive which took a day to process, now clicking in the thumbnail will open the video in google drive

if the video doesn't play when you click play, refresh and try again. weird glitch. i enabled the ability to download it if it's necessary

the demo is removed from the codebase, the quality i chose to upload here was 720p because i didn't want to hit a requirement for git-lfs even if it is an az-400 objective