implement rabbitmq and complete the architecture locally
boyyy did i procrastinate with this one and the reason why is because i had to think quite a bit about how to implement rabbitmq in a clean way as it's quite dirty compared to azure service bus

skills from many resources were used here:
- Building ASP.NET Core 3 Hosted Services and .NET Core 3 Worker Services by Steve Gordan
- RabbitMQ by Example by Stephen Haunts
- Securing Microservices in ASP.NET Core by Kevin Dockx
- RabbitMQ Custom Docker Image with Custom Configuration and Definitions by Mike Møller Nielsen [YouTube]
- How export or import RabbitMQ configuration [Article]

a custom rabbitmq image is built with all the exchanges, queues and bindings configured, i manually configured and exported this using its api

the normal guest account is used for rabbitmq

i created my own naming convention for exchanges and queues [afaik] as follows:
exchange.applicationname.event
         applicationname.event.subscriber.queue

an example:
exchange.seelanstyres.deleteaccount
         seelanstyres.deleteaccount.addressservice.queue

this satisfies my ocd as it's easy to understand what it means

explaining it further:

a publisher publishes messages to an exchange [of type fanout in this case] and queues bind to that exchange and receive a copy of those messages.
each subscriber will listen in on its dedicated queue and process the messages

a visual of a fanout exchange [sort of 😂]:

                      -> queue -> subscriber
publisher -> exchange -> queue -> subscriber
                      -> queue -> subscriber

in the system there's three events:
- deleteaccount
- updateaccount
- updatetyre

the order microservice has data related to all three so a worker is created for each
the address microservice only needs to listen for when a customer deletes their account to remove all their addresses

i tried my best to abstract a lot of the complexity of rabbitmq into the shared library and ensure there isn't much duplication, however,
there's massive duplication for the worker services, in fact, the whole thing is copy and pasted and adjusted for the situation. this is something
that i will have to address later as it's not that simple to extract a common method or extension method

security as well as distributed tracing is maintained across rabbitmq

kevin dockx' implementation for token validation is used across rabbitmq since it's the cleanest way to implement it

now that the system is complete locally, soon it'll be azure time 😁 but before that, there's a lot of optimizations to make in terms of performance and implementing redis locally too