resolve some of the duplication
this is also a precursor for implementing an efficient api versioning strategy even though i'm not going to finish it anytime soon

the models of each microservice have been extracted into its own library

here's how it'll work. in a polyrepo solution, each models library will be versioned according to the api version which is not the same thing as the codebase version. it will be published to a package feed like nuget or a private package feed created in azure artifacts. each dependent project / client will of course reference the nuget package with the version it needs

however, should the schema of the database change, it will signal a new version of the api and models, and it'll be up to automapper to map older versions of the models to the entity that's going to be persisted to the database. a new nuget package will be published and any client that needs it can simply update the package reference in the csproj files and make minimal code changes [depending on the update] to be able to use a newer version of the api

---

the reason for the model duplication in the first place was because i needed to maintain the autonomy of the microservices. putting it in the shared library is probably the worst idea ever because not all microservices need to know about the models used by other microservices

this approach maintains the autonomy of the microservices and makes it easy to version those models and the apis, allowing the system to grow seamlessly while catering for clients that don't want to update and use a newer version of the api. of course this is my project and would never keep things out of date but that's just me

this is in reference to a few az-400 objectives:
- design and implement build and release pipelines
  - design a package management implementation that uses azure artifacts, github packages, nuget, and npm
  - design and implement package feeds, including upstream sources
  - design and implement a dependency versioning strategy for code assets and packages, including semantic versioning and date-based

these objectives are from the current az-400 exam which is just reworded from the exam i took