models: align with the semver versioning strategy and introduce model wrappers
since the models are tied to the version of the api, api versioning itself will be dynamic and based on the model version

the original model will be serialized with the json serializer to its utf8 bytes representation of the model. the model wrapper has the model version as a readonly property based on a constant "currentversion"

with this approach to versioning, there's no such thing as controller duplication to deal with specific versions of the api and models at all. automapper will be tweaked to map between the models and its entity equivalent by adding or removing properties and setting defaults

all the responsibility will be placed on automapper and its mapper profiles to deal with versioning using the model version in the model wrapper

extension methods will also be created to map to and from the serialized model at the level of the api and also pass in a generic type that will be used to determine which model to deserialize to before automapper. example being BrandModel or TyreModel for the tyres model wrapper

when i came across api versioning, everyone duplicates the controllers and i won't tolerate that because it's too much of technical debt. this is the best approach to this problem that i could come up with and i can safely say that this is my very own solution to this problem

also, initially i was thinking of adding the version to the api routes but with this, it introduces another problem. ocelot would have to be modified which is pain. so to avoid that, a header will be sent with the request using the version from the models-library constant "currentversion" which nicely ties in with the version of the api to use. any clients would not need to change any code in order to work with the apis