rework how updates are published by making it more resilient to failure
my approach is quite decent here to the point where i feel like it would be quite useful to the world and may later on extract it into a new project in a new repo and publish it to nuget

anyways, here was the problem i solved. the update of course would fail to publish if the message bus wasn't available

but before i explain how i solved it, it's worth mentioning how Neil Morrissey in his course "Implementing a Data Management Strategy for an ASP.NET Core Microservices Architecture" solved this problem

what he did to solve this was create a second database to write the update to and have a second project that will poll that database and publish it to the message bus, i don't like this tactic because it's not maintainable and prefer to have the microservice itself ensure that the update is sent to the message bus. but what i did take from this is that you can write the update to the database, it doesn't really need to be a second one

my solution is a bit complex and involves:
- System.Threading.Channels for the first time
- a second dbcontext
- two background services

so System.Threading.Channels. what's its purpose and how to use them

a channel can allow you to offload processing from a request to a hosted service and provide an immediate response to the user

visually, it's pretty simple to understand

write -> channel -> read

in my solution, i offload the publishing of the update to the first background service "PublishUpdateChannelReaderBackgroundService" which is the reader for the channel

the channel is defined as Channel<(BaseMessage, string)> which reads and writes a tuple with the base message and destination to publish to

publishing of the update is first attempted and if it fails, the update is encoded to a base64url and stored in the new dbcontext, which will be migrated to the same database as the one being updated

a second background service is created to read from that dbcontext - the unpublished updates every 10 minutes - and try again. each retry is counted and logged and when an update is finally published, it's removed from the database

all the code is in the shared library and an extension method on the iservicecollection is created to:
- register that dbcontext
- add the repository
- add the channel
- add the two hosted services

in total there's 5 services, and it's reuseable and easy to integrate

on the controllers, after the basemessage is created, instead of publishing it immediately, it's written to the channel

distributed tracing isn't maintained when sending to a channel so i modified the "StartANewActivity" extension method to pass in the operation name and ensure distributed tracing is maintained

skills from Steve Gordan's course "Building ASP.NET Core 3 Hosted Services and .NET Core 3 Worker Services" were used here

this approach to this problem too is my very own one 😌

however, logging has to improved a bit and the publish-destination field has to be synchronized in all the logs