finally provide a fix for identityserver4 in docker-compose 😁
this has been a blocker for me for some time now, preventing me moving to kubernetes

skills from erik dahl's course were used here, including most of his approach. an improved way of honoring forwarded headers from reverse proxies i got from the eShopOnContainers reference architecture by microsoft

the course: Fundamentals of Docker and Kubernetes for .NET 5 Developers

the issue here was a bit complex. identityserver4 only allows dns names that resolve to 127.0.0.1 [localhost] to work for security reasons

the fix involves placing a reverse proxy in front of both the identity service and the mvc frontend to have a static dns

using powertoys 0.65.0, i added two hosts file entries to resolve to 127.0.0.1:
* id-local.seelanstyres.com [for the identity service]
* www-local.seelanstyres.com [for the mvc frontend]

this is a good demo idea so it'll be created later on

running the scripts like normal afterwards will build the containers and run them like normal with the addition of a custom nginx container

however, the login flow works but there's another flaw regarding the encryption process not working because the discovery document fails to be retrieved which is weird but a fix may come around later

a lot of small changes were also made that were not relevant for the fix:
- simplifying the database connection string
- unifying certain appsettings across the entire architecture
- pulled out the cookie policy into the shared library and changed it
- changed the "UseDocker" appsetting name to "InContainer"
- changed the location of the volumes to be placed within the orchestration folder
- removed keeping volumes for logs and secrets regarding the sqlserver container because only the data is important

and a few more