azure: instrument the solution with application insights telemetry
this is the first obvious step moving into azure and the behavior of the telemetry sent is a bit of a deal-breaker in some keys ways

azure application insights is MOSTLY FOR telemetry and not logs so this needs to be said before explaining its nuances

first, one instance is used for all the microservices with a key field being "cloud_RoleInstance" used for each application source.
a better practice would be to create an instance for each application which in a microservices architecture is a nightmare

this behavior is the first nuance

on localhost and in docker-compose, its value is the name of the machine it's running on "Shaylen" in my case but in kubernetes, it's the pod.
for a pod, it has the deployment name followed by the pod-template-hash then the unique id of the pod which is more accurate

this is the value you'd get from running 'kubectl get pods -n seelanstyres' then check the logs 'kubectl logs pod-instance -n seelanstyres'

a second nuance is a field called application_Version

on localhost, it shows the file version which seems to come directly from the assembly info but this is not that useful to me
as the informational version has the first seven characters of the git commit in it which i can use to go to github
and search for the commit id and check the code. any enrichment i added to serilog is non-existent which is a bummer

in docker-compose and kubernetes, this value is "7.0.0", which is totally off and not useful at all and may be the version of the dotnet runtime

with these nuances, looking at the end-to-end transactions [related to distributed tracing] isn't so useful as it shows the exception, fair enough but the version info
is missing so its only use for me is the drill down and get the timestamp which i can use to investigate the logs and get the commit url and check the code

---

moving forward

in a commit that seems like years ago now, "prep for centralized logging with serilog and elasticsearch", i said something about application insights and quote

    "... in the exam, for example, distributed tracing was tested using the application map of application insights and ofcourse,
    the projects have to be instrumented with application insights. i didn't do that here, not too sure how possible that is
    when the solution is in azure kubernetes services but i'll figure that out when i get there ..."

before i had any deep knowledge about kubernetes and how it works, this is what i was thinking at the time

"if a service isn't EXPOSED TO THE INTERNET, how exactly is it going to send telemetry to application insights OVER THE INTERNET ???"

i still cannot confirm this because of my cluster which i have to troubleshoot now 😐

---

the next thing i've done is publish the health check status to app insights using the "AspNetCore.HealthChecks.Publisher.ApplicationInsights" nuget package

figuring out how it works was a bit irritating because the readme on the repo doesn't tell you how it works, you have to right-click and go to the definition
in the code that adds the functionality for a full explanation and here it is:

    "For each Microsoft.Extensions.Diagnostics.HealthChecks.HealthReport published
    a new event AspNetCoreHealthCheck is sent to Application Insights with two metrics
    AspNetCoreHealthCheckStatus and AspNetCoreHealthCheckDuration indicating the
    health check status ( 1 Healthy 0 Unhealthy) and the total time the health check
    took to execute on milliseconds"

using this, i checked the logs and it falls under the "customEvents" table and this is the kusto query i created that'll be used to create an alert rule
to eventually trigger the "SystemDegradedToggler" azure function app and turn on the system-degraded state as well as notify me:

customEvents
| where name == "AspNetCoreHealthCheck"
| where customMeasurements.["AspNetCoreHealthCheckStatus"] == 0
| project
    timestamp,
    ApplicationName=customDimensions.["Assembly"],
    ApplicationVersion=application_Version,
    HealthCheckStatus=customMeasurements.["AspNetCoreHealthCheckStatus"],
    HealthCheckDuration=customMeasurements.["AspNetCoreHealthCheckDuration"]

to break it down

query the customEvents table [customEvents]
and filter for the name of the event where it's equal to "AspNetCoreHealthCheck" [| where name == "AspNetCoreHealthCheck"]
then filter for the health check status where it's equal to 0 [unhealthy] [| where customMeasurements.["AspNetCoreHealthCheckStatus"] == 0]
and only show me the following fields [| project] ...

most projected fields are renamed with a syntax of "NewName=OriginalName"
for example, ApplicationName=customDimensions.["Assembly"]

the syntax for how to check for values in json came from an example in the docs, as the "customMeasurements" and the "customDimensions" fields have its value in json

---

with the implementation of application insights, this implements and/or references quite a few exam objectives from all my exams. the az-104, az-204 and az-400.
even though this is true, i will only show the az-400 objectives because its similar to the az-104 and az-204

Develop an instrumentation strategy (5-10%)
    Design and implement logging
        - interrogate Log Analytics logs using basic Kusto (KQL) queries

    Design and implement telemetry
        - design and implement distributed tracing
        - inspect application performance indicators
        - inspect infrastructure performance indicators
        - define and measure key metrics (CPU, memory, disk, network)

Develop a Site Reliability Engineering (SRE) strategy (5-10%)
    Develop an actionable alerting strategy
        - implement alerts based on application health checks
        - develop communication mechanism to notify users of degraded systems

---

with the move to azure, i've also added the resources that i've used to study for azure.
technically, up until this point i've never even looked at them for this project so i never mentioned it.
even to implement the az-400 objectives, it's been my own skill and understanding of the objectives that i used.
so now that azure comes into play, the resources i used to learn azure also come into play