azure: implement the cloud version of the image service
this is the second step and this too, had a surprising nuance

when retrieving the blob properties after uploading the blob, the url isn't a property that you can get, i thought that might be an obvious one. the url has to be built manually

the flow to upload to an azure storage account is very boilerplate

i've seen the code sample on github and how thomas claudius huber implements it in his course "Microsoft Azure Developer: Develop Solutions with Blob Storage". however, my implementation is slightly different

the flow:

blob service client -> blob container client -> blob client -> upload

- first, you need a blob service client to connect to the storage account
- then you retrieve a blob container client from the blob service client
- thereafter, create the container if it doesn't exist
- retrieve a blob client from the blob container client and provide a unique name for the blob
- upload the blob using an open file stream to the storage account

the url is built using the properties of the blob service client, the blob container client, and the blob name

$"https://{blobServiceClient.AccountName}.blob.core.windows.net/{blobContainerClient.Name}/{blobName}"

---

the storage account kind i chose for the trials here is a premium lrs blockblobstorage

lrs being "locally-redundant storage", the storage account is deployed to south africa north and no zrs [zone redundant storage] option is available

so why not general-purpose v2 ??

i'm only using a storage account for the images, nothing else, not even logs, so everything else is pointless

however, i could change my mind later because of the replication options of the general purpose v2

with that, there are quite a few options

lrs - locally-redundant storage - three copies in one data center
zrs - zone-redundant storage - one copy in each availability zone in the primary region
grs - geo-redundant storage - lrs in the primary with lrs in the paired region
gzrs - geo-zone-redundant storage - zrs in the primary region with lrs in the paired region

ra-grs - grs with read access to the paired region
ra-gzrs - gzrs with read access to the paired region

so, if i want more redundancy with the premium blockblobstorage, i have to create a second storage account in another region and enable blob replication to that account. a bit annoying

---

security of the account will drastically improve when i run the entire solution in azure, but for now, i just wanted to test the code

---

this implements an az-204 exam objective:

Develop for Azure storage (15–20%)
    Develop solutions that use blob storage
        - Perform operations on data by using the appropriate SDK