implement functionality to delete addresses
seems like a pretty obvious feature that's been missing, right?

here's why:

initially when the database was one giant monolith, if an address was deleted, it would cascade and delete the orders too, and this is not the behavior i wanted, so it was never implemented, even in college

the original database structure in sql was removed with the commit "remove the sql script of the database" back in august 2022

also, another weird realization occurred, i never checked my spelling of the word "preferred" and it was wrong with the spelling of "preffered" for a while. somehow this wasn't a consistent issue throughout the solution. weird

deleting an address will not be an event that'll be raised in the architecture

orders need to remain intact and true to the original purchase

this is also why updating addresses doesn't exist and will never be implemented