slightly improve on the retry unpublished updates worker
improvements:
- when getting the 'IUnpublishedUpdateRepository' from the service provider, use GetRequiredService instead, this ensures that it's not null otherwise an exception will be thrown, this should never happen anyway because all the services are registered together. this removes the null-forgiving operator in this scenario
- when decoding and deserializing the message from the unpublished update, coalesce it with a throw new InvalidOperationException, this will never happen so it's safe to do so. this also aids in removing all the null-forgiving operators around the message
- Thread.Sleep now uses the TimeSpan.FromMinutes(10). this was done to improve readability which should - and will - become a change everywhere the milliseconds were calculated

the reason why i touched this is because i noticed that the wiki page around this topic has been outdated for some time, it'll be updated with this change. also, i recognized too much of null-forgiving operators [which is in the entire solution] while there are better ways of dealing with these, so i had to adjust it

one of the reasons why i haven't taken this out into a new repo is regarding the creation time. it's a DateTime, which will be decided based on the server the solution is on while completely disregarding the time zone information. DateTimeOffset is a better option, however, i'm not too sure if sql server supports this type out of the box. from the course 'Globalization and Internationalization in .NET 6' by Filip Ekberg on pluralsight, DateTimeOffset was converted to a string using the 'O' format [iso 8601 culture-invariant roundtrip time], so the offset is persisted. then if the server has it wrong, you can convert back to a DateTimeOffset and apply the ToLocalTime() to get the right time and time zone

this solution isn't globalized though. there's a lot of work to do that [especiallly around the dates, supported cultures, number formats and localization of the views], but i may not do so with this project, and may build something else in the distance future from scratch [like i've mentioned in the wiki] that follows a lot of this architecture and devops, while including better practices around globalization and internationalization

with regards to this library however, i might take this issue into account and implement it here before pulling it out to see if it works the way i expect it to, so there's no problems