logs: perform more log reduction and cleanup
also, since one of the code smells was the repetition of the image url for tyres, i resolved this and added a new migration to see a difference and the result wasn't what i wanted. this removed all the old data and re-added them, and this was due to the fact that i used "Guid.NewGuid()" for the id of the tyre instead of defining it. so to solve this, i just went into the migration and pulled the generated ids and added it to the seed data in the dbcontext. this resolved that issue

the repetition of retrieving the customerId from the claims has also been resolved

[skip ci]