rename the mvcbff project to webbff
this is a MASSIVE breaking change in this solution

luckily, i never officially release v1.0.0 of this solution, so technically, i can still make changes like this

the reasoning behind it is because i don't want the name to be tied to mvc. a generic name like "webbff" opens up the possibility of multiple web frontends

this will render the name "mvcbff" technically incorrect

essentially, all databases were scraped and re-migrated

it's only now that i also noticed that since i decided to make a poco out of the elasticsearch logging sink configuration section, that it is required when starting up the applications. this broke the database migration script. this is now resolved, also in the release pipeline

funny enough, after i had taken down the solution from azure, i decided to start with with removal of using iconfiguration directly. if i had kept the solution up in azure, i would have noticed that the release pipeline broke, and i would've fixed it then

[skip ci]